requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-jwt-extended
python-dotenv
bcrypt
Flask-JWT-Extended